Update global CSS variables and add custom global styles
This commit updates the global CSS variables in the `globals.css` file to change the background and foreground colors. It also adds custom global styles for the `html` and `body` elements. Additionally, it includes new `@font-face` rules for the 'BrandonGrotesque' and 'Calibri' fonts. Finally, it adds a new utility class `.text-balance` in the `@layer utilities` section.

Update Tailwind CSS configuration

This commit updates the `tailwind.config.ts` file to include additional `textShadow` options for different text sizes. It also updates the `colors` section to use the updated global CSS variables for the background and foreground colors. Finally, it adds new `fontFamily` options for the 'BrandonGrotesque' and 'Calibri' fonts.